SDIL-829 SDIL-831 increase max litreage and minor fixes

+++ app/sdil/models/Litreage.scala
@@ -18,7 +18,7 @@ package sdil.models
 
 import play.api.libs.json.{Json, OFormat}
 
-case class Litreage(atLowRate: Int, atHighRate: Int)
+case class Litreage(atLowRate: Long, atHighRate: Long)
 
 object Litreage {
   implicit val format: OFormat[Litreage] = Json.format[Litreage]
